requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
docopt==0.6.2
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pipreqs==0.4.11
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
yarg==0.1.9
